+++ assets/js/atomic/blocks/product-elements/category-list/block.tsx
@@ -7,6 +7,7 @@ import {
 	useInnerBlockLayoutContext,
 	useProductDataContext,
 } from '@woocommerce/shared-context';
+import { useColorProps, useTypographyProps } from '@woocommerce/base-hooks';
 import { isEmpty } from 'lodash';
 import { withProductDataContext } from '@woocommerce/shared-hocs';
 import { HTMLAttributes } from 'react';
@@ -16,10 +17,6 @@ import { HTMLAttributes } from 'react';
  */
 import './style.scss';
 import { Attributes } from './types';
-import {
-	useColorProps,
-	useTypographyProps,
-} from '../../../../hooks/style-attributes';
 
 type Props = Attributes & HTMLAttributes< HTMLDivElement >;
