+++ assets/js/atomic/blocks/product-elements/title/block.tsx
@@ -11,17 +11,17 @@ import { isFeaturePluginBuild } from '@woocommerce/block-settings';
 import { withProductDataContext } from '@woocommerce/shared-hocs';
 import ProductName from '@woocommerce/base-components/product-name';
 import { useStoreEvents } from '@woocommerce/base-context/hooks';
+import {
+	useSpacingProps,
+	useTypographyProps,
+	useColorProps,
+} from '@woocommerce/base-hooks';
 
 /**
  * Internal dependencies
  */
 import './style.scss';
 import { Attributes } from './types';
-import {
-	useSpacingProps,
-	useTypographyProps,
-	useColorProps,
-} from '../../../../hooks/style-attributes';
 
 type Props = Attributes & HTMLAttributes< HTMLDivElement >;
